Separate secrets into multiple files
